# gildongE/src/main/java/com/gildong/gildongE/dto/UserCreateRequest.java
package com.gildong.gildongE.dto;

public class UserCreateRequest {
    private String userName;

    public String getUserName() {return userName;}
    public void setUserName(String userName) {this.userName = userName;}
}