# gildongE/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: GildongE API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local server

tags:
  - name: Users
    description: 사용자 관련 API
  - name: DrivingPatterns
    description: 운전 패턴 관련 API
  - name: Consumables
    description: 소모품 교체 예정일 관련 API

paths:
  /api/users:
    post:
      tags: [Users]
      summary: 신규 사용자 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: 잘못된 요청 (유효성 실패)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/users/{userId}:
    get:
      tags: [Users]
      summary: 사용자 조회
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 사용자 고유 ID
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Users]
      summary: 사용자 이름 수정
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 사용자 고유 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/patterns:
    post:
      tags: [DrivingPatterns]
      summary: 운전 패턴 기록 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrivingPatternRequest"
      responses:
        "201":
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrivingPatternResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/patterns/user/{userId}:
    get:
      tags: [DrivingPatterns]
      summary: 사용자별 운전 패턴 목록 조회
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 사용자 고유 ID
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DrivingPatternResponse"
        "404":
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/consumables:
    post:
      tags: [Consumables]
      summary: 소모품 교체 예정일 저장
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumableRequest"
      responses:
        "201":
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumableResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/consumables/user/{userId}:
    get:
      tags: [Consumables]
      summary: 사용자별 소모품 내역 조회
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 사용자 고유 ID
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsumableResponse"
        "404":
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 에러 메시지

    UserCreateRequest:
      type: object
      required:
        - userName
      properties:
        userName:
          type: string
          description: 실명

    UserUpdateRequest:
      type: object
      required:
        - userName
      properties:
        userName:
          type: string
          description: 변경할 실명

    UserResponse:
      type: object
      properties:
        id:
          type: string
          description: 사용자 고유 ID
        userName:
          type: string
          description: 실명
        avgDrivingScore:
          type: number
          format: float
          description: 평균 운전 점수
        createdAt:
          type: string
          format: date-time
          description: 생성 시각

    DrivingPatternRequest:
      type: object
      required:
        - userId
        - drivingScore
      properties:
        userId:
          type: string
          description: 사용자 고유 ID
        drivingScore:
          type: number
          format: float
          description: 운전 점수

    DrivingPatternResponse:
      type: object
      properties:
        id:
          type: string
          description: 기록 고유 ID
        userId:
          type: string
          description: 사용자 고유 ID
        drivingScore:
          type: number
          format: float
          description: 운전 점수
        recordedAt:
          type: string
          format: date-time
          description: 기록 시각

    ConsumableRequest:
      type: object
      required:
        - userId
        - carModel
        - carNumber
      properties:
        userId:
          type: string
          description: 사용자 고유 ID
        carModel:
          type: string
          description: 차량 모델명
        carNumber:
          type: string
          description: 차량 번호
        engineOilDate:
          type: string
          pattern: '^[0-9]{8}$'
          description: 엔진 오일 교체 예정일 (YYYYMMDD)
        batteryDate:
          type: string
          pattern: '^[0-9]{8}$'
          nullable: true
          description: 배터리 교체 예정일 (YYYYMMDD)
        coolantDate:
          type: string
          pattern: '^[0-9]{8}$'
          nullable: true
          description: 부동액 교체 예정일 (YYYYMMDD)
        transmissionOilDate:
          type: string
          pattern: '^[0-9]{8}$'
          nullable: true
          description: 변속기 오일 교체 예정일 (YYYYMMDD)
        brakeOilDate:
          type: string
          pattern: '^[0-9]{8}$'
          nullable: true
          description: 브레이크 오일 교체 예정일 (YYYYMMDD)
        washerFluidDate:
          type: string
          pattern: '^[0-9]{8}$'
          nullable: true
          description: 워셔 액 교체 예정일 (YYYYMMDD)

    ConsumableResponse:
      type: object
      properties:
        id:
          type: string
          description: 소모품 내역 고유 ID
        userId:
          type: string
        carModel:
          type: string
        carNumber:
          type: string
        engineOilDate:
          type: string
        batteryDate:
          type: string
        coolantDate:
          type: string
        transmissionOilDate:
          type: string
        brakeOilDate:
          type: string
        washerFluidDate:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time